eliminate jQuery.ajax() calls
and some other changes, but that's the main difference. Using
fetch() instead will break the client for legacy broswers like
IE, but let's not rush to fix that, there are a number of possible
fixes: https://www.javascriptstuff.com/ajax-libraries/ and we should
pick one deliberately.

I'm assuming we'll use something like Vue, rather than jQuery, for the
UI, and Vue doesn't handle ajax, hence this change.

Leo build: 20180217163638